Fast settings... go vruuuum
Added this 3dbenchy with my fastest settings. It doesn't look nice, but it prints. You can normally use printer.cfg I have on GitHub and get the fastest prints. It's mostly in slicer settings. You may have problem with cooling or klipper shuting down, because of lack of the compute resources. If you want to get the fastest prints, you will need to run klipper from notebook or desktop.